Add mocking service
This is the first working prototype, being able to mock a specified
interface. A return must be manually added at the end of each function
in order to have a compiling mock.
Functions can be overwritten at the mock initialization.

The current architecture is a bit messy, it will be reorganized after
the basic functionality is in place.